FEATURE: Add native ssh implementation
This patch adds a new  'ssh_type' === 'native' to deeployer that uses local ssh and scp commands instead of
php-ssh implementations.

Currently the server-options host, port, user and identity_file.private_key are supported. Since native ssh is used the settings of your ~/.ssh/config file apply to the connections.

The use of username + password is not implemented yet in favor of key authentication.